went back and edited Jared's code
I realized that Jared was using the same libraries I was using.

After researching and talking with him, it seems that this is
the best library out there for the job

The only downside is that the image is not completely crisp
and that the 'screenshot' may be browser/screen size dependent.

I changed the layout of the PDF to be portrait instead
of landscape, and adjusted all of the screenshots to fit
in the new page layout.

The pages look more crisp overall. I also managed to fit
more information on each page, dynamically.